snapshotter: do not duplicate new snapshotter code
