+++ contrib/nydus-snapshotter/pkg/filesystem/stargz/digest.go
