+++ contrib/nydus-snapshotter/pkg/filesystem/stargz/digest_test.go
@@ -16,10 +16,9 @@ func Test_digest_Sha256(t *testing.T) {
 	}{
 		{
 			name: "testdigest",
-			d: digest("sha256:12345"),
+			d:    digest("sha256:12345"),
 			want: "12345",
 		},
-
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
